Added Home Button and Different Input Forms
I added a home icon and button for the home link in the top left hand
corner.

Also added X and Y value inputs forms with placeholder text in the form
to help the user know what kind of data they should be typing there.

Need to make the inputs mobile friendly for viewing and remove the
other input forms.