Adding Nav Bar and dropdown menu
Added nave bar for home link  and Graph it button to go.

the dropdown menu on the top left is for all the habits being tracked
and a place for the “Clear Habits” function aka refresh.

When clicked the text turns blue. In the drop down, clicking makes the
text turn blue and the section to be highlighted with a darker
background.

The home link works. I need help making the graph it button to actually
graph it and well as the habits buttons to work as well.